Created Leveldb and Pebble option functions
