+++ go.mod
@@ -3,8 +3,9 @@ module github.com/rotationalio/honu
 go 1.16
 
 require (
+	github.com/cockroachdb/pebble v0.0.0-20211021161301-9106d5d2238f
 	github.com/golang/protobuf v1.5.2
-	github.com/stretchr/testify v1.5.1
+	github.com/stretchr/testify v1.6.1
 	github.com/syndtr/goleveldb v1.0.0
 	google.golang.org/grpc v1.39.0
 	google.golang.org/protobuf v1.27.1
